getnow(): get rid of this, just use time(); remove utAppendTime()
Implement appropriate workarounds using struct tm/localtime() for
those functions that really did need the "now[]" array for time
formatting.

For the only uses of utAppendTime is doing an "i time" (info) command,
so just remove it.  Nobody in the 21st century needs to ask for the time
on conquest anyway.