remove ratfor.c, no longer needed
ratfor.c was a file created during the initial port from RATFOR
(RATional FORtran) to C.  It implemented certain functionality that
seemd to be intrinsic to RATFOR.

Now that we've removed the need for or re-implemented elsewhere
those RATFOR functions, we can finally remove the ratfor.c file
altogether.  Only took about 20 years. :)