MAXSHIPS work, (0-based Ships[]) and messaging rework
This large commit reworks the Ships[] common block array to be
0-based.

As a result, some rework of the messaging protocol and common block
message representation was required, as well as all messaging related
code.

This involved a lot more changes that I expected, but will be better
in the end.