remove SPECIAL_BASE and DEATHSTAR
The deathstar was never implemented in the original, and if we wanted to
add new "eaters", we should do it via the doomsday protocol/CB via a
new "eaterType" member.  I'm thinking "Nomad" and "Space Amoeba"
in the future perhaps :)

Assign DISPLAY_DOOMSDAY to -10 which should be safe with new code
(hence no need for SPECIAL_BASE hacks.