rework detection and specification of stationary planets
Previously, stationary planets were hardcoded and there was various
logic to detect this and do the right thing.

Additionally, CQI had the notion that if a planet had a primary (body
the planet orbits) that was equal to itself, then that planet would be
considered "stationary".

Murisak, hardcoded at the center of the universe was used as a backup
hardcoded primary when one was not specified or found.

Now, a primary must always be specified, and there is no hardcoded
fallback to Murisak.  Instead the fallback is for the planet to orbit
itself (become stationary).

Various logic fixes to detect this case now, and provide a
non-hardcoded fallback.  Additionally, this fixed some issues
introduced when the Planets[] array was made 0-based.  A primary == 0
is perfectly valid now.