implement cqi Planets[] core, homeplanet, rework relevant logic
This commit removes the last vestiges of the hardcoded core planets
(planets needed to take the universe), hardcoded team "suns" and
hardcoded team "home" planet(s).

Previously, each team had to have exactly 3 home planets, and they had
to be specific planets.

Now with the cqi Planets[] init system fully operational, the only
requirement now is that each team must have at least one "home"
planet.  More is fine, but at least one is required.  Any home planet
owned by a team is a candidate for robotic defense when attacked.

In addition, the cqi "core" functionality was completed.  A "core"
planet is a planet that must be taken in order to conquer the
universe.  Previously, that was always the first 25 classd or classm
planets.

Now, whether a planet must be conquered in order to conquer the
universe is determined soley by a planet's "core" flag being set.