history: rework and finish protocol/CB changes started long ago
Finish implementing the history changes.  Previously, History[] was
intimately tied to the Users[] table.  This made is difficult for
detecting a variety of things, such as when a user resigns.  All trace
of that resignation would disappear from the History table.

Now the username is stored along with the history entry, so that if a
user resigns, this can be detected.  Resigned usernames are printed in
cyan on the history list.

Also, related Users[] packets are no longer required to decode and
record a history entry.  User numbers are still stored there for
conquestd book keeping.  User numbers equal to -1 mark resigned users.