add the team character to the planetlist for home planets
This makes it possible to know which planets are a team's
"homeplanets".  Homeplanets of a team must be conquered in order to be
able to enter from that teams primary system in addition to your own.

It is listed in the planet list as a character representing the team
('F' = federation, etc) along with the planet's core status ('+'),
before the planet's name.