rework SwitchTeams server option, remove 'multiple' knowlege from clients
For some reason, the SwitchTeams server option was implemented as a
User option, when it's not.  Rework so that the User option is
ignored, and server flags (which include SwitchTeams) is honored.
Several of these User options will be going away, or refactored in
upcoming commits.

Removed all knowlege of 'multiples' Users from the clients.  These
Users are reserved for the Team users ("Federation", "Klingon", etc)
so that their robots can fly more than one ship at a time.  This isn't
used for "normal" users like everyone else, which are only allowed to
fly one ship at a time.

Eventually, the "multiple" stuff will go away completely in favor of a
special "team user" flag for the relevant users that won't affect
anybody else.