implement userTypes_t to represent various user types
This is used to distinguish between internal "builtin" users like
"Federation", "Klingon", etc.  This makes it possible to prevent
someone from being able to login to a builtin user, but still be able
to choose a login name like "Klingon" without there being any
confusion in the code.

Builtin users are marked by an '_' preceeding their alias in the user
list.

We should be able to remove all traces of the "multiple" code/options
now since we will use the user type to determine whether a user has a ship
occupancy limit.