fix most remaining .cqr playback issues with new Ship[]/Planets[] and messaging
This should now allow proper playback of recorded games that used the
v6 protocol.  Ship and Planets numbers are compensated for properly.

Messaging playback fixed to translate from the old single integer
representation of the to/from to the new message addressing format.
Modified the playback clients to correct for the new ship numbers when
(w)atching a ship, and cycling through them.

There is one issue with the old planet/ship indexing that cannot (or
won't) be fixed.  Some message texts include a ship number, such as
"All hail the liberating armies. Thanks K3!".  In these cases the ship
number will be off by one, as there is no attempt to correct the
actual message bodies.  That would be insane :)