convert users of strncpy() to utStrncpy() which always does the right thing
It will always ensure that the buffer is 0 terminated.

This should be safer and less confusing.  It would also appear that in
my early Conquest porting efforts, I did not really know how to use
strncpy() properly in some cases, so those have been fixed too.

Also fixed a problem with the hud lasttarget not being initialized
with the planet name when scanning for a planet in the GL client.