fixup the MAXHOSTNAME + 10 crap, define new MAXPORTNAME and change MAXHOSTNAME
According to RFC 1035 the maximum length of a fully qualified domain
specification is 255, not 64, so set it correctly.  I should probably
be querying this stuff from the network stack (if possible).

Instead of doing a "+ 10" to compensate in some areas for the port
specification, create new MAXPORTNAME and use that where appropriate.