cb.[hc]: more rework, get rid of hardcoded SIZEOF_COMMONBLOCK
Some cosmetic, naming, and other minor changes.

Add a new cbGetSize() to determine the required size to hold the
common block as configured.  Use this to do allocations and mapping.
This also ensures that we don't need to over-allocate space as we did
before -- we allocate only what we need now.  Get rid of
SIZEOF_COMMONBLOCK macro and old 'comsize()' conqunix.c function.

Add the ability to unmap the common block - cbUnmap() and
cbUnmapLocal().

Previously, when we found a common block who's size was different than
expected, we would simply delete the common block file (conquest.cb)
and create a new zero-filled one of the correct size.  Now, if this
happens, the original file will be renamed instead (with the unix time
code appended) so it isn't lost in case of an operator accident.

Right now, we are still using hardcoded defaults (60 planets, 20
ships, etc) so it's not an issue.  But soon, these and other values
will come from the conqinitrc file's "global {}" section. When that
happens, changing any of those values will change the required CB size
and hence require a complete re-init of the universe.