finish implementing the 'size' parameter of the planets
As usual, this turned out to be a bigger deal than anticipated.

Previously in Conquest, planet sizes (and suns) were hardcoded to a
size of 300 CU's.  The orbit distance was the same, and you started
taking damage from a hostile object (planet/sun), or scanning it, at
500 CU's.

Some time ago, CQI and conoper were modified to allow specifying the
size for a planet.

The above limits (300/500) were still always used, even though it was
now possible to draw them at different sizes.

There was also an OBJ_PRESCALE variable, set to 1.3, being used in the
GL client to scale up all of the objects by that amount, so things
didn't look too tiny.  That was a mistake.

This all needed to be changed.  Prescaling should be applied to a
texture to only compensate for issues in the texture itself, and for
display purposes *only*.  And not everything needs to be prescaled.

For example, the "star" texture is mostly made up of the "glow"
surrounding it.  The actual "core" of the star is small in comparison.
For this reason, we would want to draw it a little larger so that if
you orbit it, you will be inside the glow, not outside.  Now we can do
that. :)

This prescale set on a texture is only used for rendering.  Currently
it is only used for ships, planets, and the Doomsday machine.

The protocol has been modified to send the planet size to the client
now.

The planet's size is now never prescaled (though the texture used to
draw it can now be), and the range at which you can scan or receive
damage from a planet depends on it's "true" size now, regardless of
how it's being drawn.

conqinitrc was modified to make the "jupiter" planets at twice the
size of the other planets (600 CU's), as seems fitting for a gas
giant.  Suns were already set at 1500 CU's and everything else is
still the default 300.

The 00_plantexrc.trc file was modified to add appropriate prescaling
to:

 - the stars to compensate for the glow (1.7x actual size)

 - "uranus" planets, to compensate for the ring (2x actual size)

The texturesrc file was modified to add appropriate prescaling to:

 - ships (1.3x actual size)

 - doomsday (1.3x actual size)

Both the ships and doomsday should look the same as the previous
version.

The conqinitrc file is now unconditionally overwritten on a make
install (it used to check to see if it existed first).  In a future
commit this will be handled differently, perhaps by allowing
specification of that file on the conquestd command line, so if you
want to make a custom one, you can still use it without it being
overwritten by a new version.

NOTE: all of this will be meaningless to the curses client -
everything will appear as it normally does (the same size), but you
may wonder why you seem to be taking damage from a star you appear to
be far away from.

It's getting time to just retire the curses client.