fix bombing bug that occasionally caused bombing to stop
During a play test with a friend, it was noticed that bombing would be
stopped by the server when trying to bomb a planet.  This turned out
to be do to accidentally sending a ping packet to the server during
one of the bombing states (S_BOMB).  This would cause the server to
stop bombing, making the player very sad.

In future commits, the whole PING and KEEPALIVE mechanism will be
reworked so this kind of thing won't be an issue anymore, and the
special checks for this issue can go away.