Remove the curses client (conquest) and rename conquestgl to conquest
This commit removes the conquest curses client.  Anyone playing on
the curses client against a player using the GL client (formerly
conquestgl) has been at a disadvantage for a long time.

With some of the new changes to planet rendering (size) and the like,
it's only going to get way worse.  Besides, I'm tired of trying to
maintain it :)

"conquestgl", the GL client, has been renamed to "conquest" and some
no longer used code/files removed.  Documentation and man pages have
been updated.

There will be further removal of dead functions and the like in the
future as I stumble across them.