Convert to building with a C++ compiler, using the C++11 standard
The .c files have been renamed to .cxx.  The header files have not
been renamed, and probably won't be unless there is a specific need.
They are never exported to 3rd parties and are only used within
Conquest.

We now use a C++ compiler to build Conquest, specifically, you need a
compiler that understands the C++11 standard - this should include
clang (3.2+) and gcc (4.8+).

Although we have not actually converted Conquest into C++, it would be
advantageous to be able to start moving in that direction since C++
can buy you a lot in terms of available containers and classes.

To do this, several fixes had to be made to get the source to compile.

There are a *ton* of warnings regarding the non-constness of string
literals, which are used all over the place in Conquest.

For now, we disable those warnings (-Wno-writable-strings), though we
will attempt to fix these as we go along for correctness.

I have tested and run builds using g++ 5.4 and clang++ 3.8.

The constrat executable also showed some new linker issues.  It might
have only accidentally been linking correctly in the past.  Until
those are properly resolved, conqstrat can no longer be used to
directly initialize the robot strategy tables in the common block (the
-U option has been disabled).  We'll deal with this later, though I
doubt anyone has actually ever used constrat that way but me for
testing, so don't hold your breath :)

In addition, to resolve the linker issues, conqstrat is never linked
to any external conquest libraries anymore.  This mostly has to do
with collisions with the two yacc/lex parsers - the one in constrat
for building the robot strategy tables, and the other in conqinit for
universe/game initialization.