GL: fix a problem in drawShip that didn't account for ship size
drawShip() assumed that all ships were always drawn at the same size
and did not properly account for a ship's actual size nor the proper
prescale defined for each ship.

In addition, set the prescale on the Orion Cruiser to 1.7 to account
for the texture, so it looks larger than the scout and destroyer -- as
it should.