Create a new circular tactical grid to replace existing grid textures
The problem with the former per-team tactical grids is that, while
they looked pretty, they were completely useless in actual combat
since there was no way to scale them properly to Conquest's coordinate
system.  They were just textures that were simply overlaid in the LR
scan (only, if enabled).

Now we do something much simpler, and IMO, much more useful.  We draw
10 concentric circles around your ship, spaced 1000 CU's apart.  The
first circle is red (1000 CU's away, corresponding to your phaser
range).  The second is yellow (2000 CU's away), and the third is green
(3000 CU's away).

Rings 4-9 are drawn in grey, spaced 1000 CU's apart and the last ring
(corresponding to 10,000 CU's) is drawn in cyan.

We re-use the options for the original grid (enable/disable and
opacity/brightness) for the new grid now.

It is also shown in both short and long range displays now, properly
scaled.

ALT-G can be used to immediately toggle the grid on/off during play,
just like the old versions, except as mentioned above, the grid is
always displayed in both LR and SR scans if enabled.