doomsday: rework CB and protocol representation, drawing
The doomsday CB struct has been modified to include a new flags field
(removing the 'status' field.).  In addition a new eaterType field was
added in case we want to add other eater types in the future.

The device is now rendered in both SR and LR views.  The anti-proton
beam is now displayed when the 'attacking' flag bit is set.  Before,
we would try to guess when the machine was attacking since the
appropriate flag was not available.

Various doomsday constants were moved out of conqdef.h and into
doomsday.h. The only two "status" values, DS_ON and DS_OFF are gone
and now represented by a flag (DOOM_F_LIVE).  Others were just deleted
due to not being used or useful anymore.