towing: fix bugs with ship slot 0, and rework
During some play testing, it was discovered that any towing/untowing
involving ship slot 0 was pretty broken, due to the recent 0-based
cbShips[] changes.

Rework and implement as a pair of flags SHIP_F_TOWING and
SHIP_F_TOWEDBY, while keeping the old .towedby and .towing ship
variables -- now used to store the relevant ship number only, and no
longer using it to store towed/towedby status itself.  That's what the
new flags are for :)

In theory it would now be possible to be able to tow and be towed at
the same time, however this is not implented... yet.

Here we just want to get things working properly the way they used to.

CB and protocol changed of course.