towing: rework, and allow tow chaining
Reworked the towing code so that a ship can tow a ship while also
being towed.  Previously a ship could only be towed by someone, or
towing someone, never both at the same time.  Not anymore.

This allows ships to chain themselves together.  One limitation: you
cannot tow a ship that is currently towing you.  That would be insane
:) Other than that, have fun.

In the future there may need to be some work to aggregate the
energy/heat consumption on a string of ships so the ship pulling all
of the others feels that "heat", as it were.  For now the fuel/heat
logic is the same as if you were only towing a single ship.

The commands for towing changed a little.  Like the 'C' cloaking
command which enables and disables cloaking, 't' is used to both tow
and untow a ship.  Use 't' to tow a ship as usual, and then to let go
of it, just hit 't' again.

The 'u' command is now only used when trying to untractor from someone
else who is towing you.

Now, if we only had a rendered tractor beam effect...