conquestd: re-arrange the order of some calls, don't load textures
We don't need to load the textures ffor the server -- it doesn't need
them.

Also, load the conqinitrc file BEFORE mapping the universe.  This will
be needed for future work, since the final common block configuration
will depend on the limits specified in conqinitrc.

The client will also eventually need to query these from the server
prior to mapping it's own local CB.