Dynamic CB part 1: starting the work
Created a cbMaxLimits class that will store the limits of the various
important arrays in the common block (CB).  It is global and
initialized in cqiLoadRC when successfully reading the conqinitrc file
and it's global {} section.

This will eventually allow us to remove several old hardcoded limits on
them.

For now, he client crashes in cqsInitSound due to references to
cbTeams[], and no CB is mapped at the time.  Will be fixed in a later
commit.