Dynamic CB part 2: cqsound and cqiFindPlanet()
Fix up cqsound so it does not need access to the CB which won't be
mapped yet.

Get rid of call to cqiFindPlanet() in _get_glplanet_info() (GL.c) so
that a NULL cqiGlobal isn't a problem.  It is NULL because we do not
want to read in the conqinitrc file in the client.

This was only used to detect if a texname was specified for a planet.
Now we do not support that at all (it was not being used anyway).