protocol: use a bitmap for srpwar states in the protocol (only)
Currently srpwar states are encoded as booleans in the ship structure.
This list was being sent with each spShip_t packet, 60 bytes worth, 1
byte for each planet (limited to 60 in old protocol).  With the
possibility of supporting up to 256 planets in the new protocol, this
could have meant 256 bytes for each ship which is too much of a waste.

So now we encode them in a bit representation in 8 uint32_t's.  This
is 32 bytes, and represents the srpwar states for all possible planets
(256 in current protocol).

Protocol changed, of course.