cbLimits: remove MAXPLANETS hardcoded limit, the first of several to come
This marks the end of the prep work to make it possible to remove all
remaining dependencies on the hardcoded MAXPLANETS (and others) limit.

This patch starts by replacing MAXPLANETS first.

Now we use cbLimits.maxPlanets() which no longer ties the number of
supported planets to 60 (only).

I'd like to thank the Academy, my friends, and my family who supported
me throughout these troubles.

There's more to come: MAXSHIPS, MAXMESSAGES, MAXHISTLOG, MAXTORPS, and
MAXUSERS are the remaining globals due to be converted to cbLimits
usage in followup commits.