cbLimits: remove MAXSHIPS hardcoded limit
This was the last and largest one.  Now all important dynamic state is
configurable (via cbLimits) now, without any more hardcoded limits on
the sizes of important things like ships and planets :)

We will also convert MAXTORPS as well, though we will hardcode it's
cbLimit to 9, since we aren't ready for that yet (shiptypes {}
section handling would need to be completed).

We should now have enough information to be able to have the server
send the client the cbLimits values of the server (via the spHello
packet), and the client will then be able to recreate it's version of
the universe on the fly as configured by the server.