protocol: send cbLimits to clients so they can build a universe
This commit adds the cbLimits data to the spHello packet so the client
knows how to make a compatible universe.

Also fixed up some issues where the client would try to call
cqiInitPlanets() and other things (semLock/semUnlock) it should not
call.

Added a new server flag for the spHello packet so that access can be
denied to a client in the future, if needed.

Recording/Playback needs fixing - the cbLimits data still has to be
recorded (in the file header), so the playback client can recreate the
right universe.  This isn't done yet.

In addition, some of the display screens will need fixing for the
possibility of being able to display more than one screen full of data
(like nShipl, nHist, etc).

But once those are dealt with, then the major work for supporting the
globals {} section will be complete!