cqi: remove "default" builtin data (initdata.h and sounddata.h)
Previously, when there was a failure to load conqinitrc, texturesrc,
or soundrc, built in defaults would be used instead that were
specified in the header files initdata.h and soundata.h.

Now, we require those 3 files to exist and parse correctly - failure
is considered fatal now.  You will need to correct the problem before
continuing to try to run anything.

While in the neighborhood, added '%option yylineno' to the CQI parser
and modified yyerror() so that syntax errors can give you a general
clue as to where during the parsing of a file that things went
horribly wrong.