isagod: add a check to see if your group is same as CONQUEST_GROUP
This is useful for those installing conquest in their own home
directories and using their own group id as the conquest group id due
to a lack of admin privileges (such as on a shared machine).

While there (isagod()), rewrote to be more efficient.  Over the years,
this function was suffering from some severe bitrot and needed some
cleaning.