cqiLoadRC: allow local (.local) variants of the main 3 cqi config files
This commit allows a user to create a local version of the 3 main CQI
config files - conqinitrc, texturesrc, and soundrc.  Simply make a
copy of the relevant file with a ".local" extension in the same
directory (<prefix>/etc/conquest/), and cqiLoadRC() will load that
file instead.

These 3 files are overwritten on a make install or package update, so
this makes it possible to make local customizations to these files
that will survive such things.