utLog(): Add special format attribute to declaration
As a "printf" style var args function, declare it as such so that the
compiler can detect mismatched types and/or missing arguments
according to the format specifier.  This is helpful to prevent
potential stack corruption on a screwup.

Fixed up some erroneous uses pointed out by the compiler.

This was tested with gcc 5.2 and clang 3.8 without problems.