nShipl: add ability to list all ships, not just the old hardcoded 20
With this commit, the work started a long time ago to get a
configurable server in terms of planets, number of ships, etc in the
CQI global{} section of conqinitrc, is complete.  The clients even
work too - now downloading the limits from the server and creating
their own dynamic CB (universe) to match.

Next up will be work to redo UDP packet handling and other protocol
cleanups.  Then hopefully we are close to a new release.  After 8
years.  Ahem, sorry.