Improve client packet handling, redo ping/keepalive handling
Redo the way pings and keepalive's (UDP) are handled in the server.
Now these packets are intercepted and handled in the packet reader
itself, so no one else needs to worry about them.  They are never kept
in the ring buffers after being read, so there should be no more need
to protect against that in the various nodes.

Now pings are sent by renderNode() as are keepalives, so no one needs
to care about remembering to send them.  It's all handled
"automatically".  This has been a long time coming - the original
implementation was just hacks upon hacks, causing nodes to have to be
very careful about pings and ping responses interfering with their work.
No more of that crap.  Any code in the nodes dealing with that have
been removed.

Changed the node return type into an ENUM instead of an int, and added
NODE_OK_NO_PKTPROC.

All packet handling is now done in renderNode(), except for certain
nodes (like nAuth) that do their own packet handling (most nodes do
not).

For those that do their own packet handling, the node's Display()
method can be used to return an indicator (NODE_OK_NO_PKTPROC) to the
node renderer so it will not attempt to do it's normal packet
handling.

Only a few nodes actually need to do their own packet handling, and
perhaps we can eliminate the need for them to do that at all
eventually.

Pings are now enabled starting at the menu node, rather than in the
CP node, and run until client exit.