Redo the way the emergency pulse messages are shown
Previously, the various emergency pulse messages ("Engines Critical",
etc) would be shown centered horizontally in the viewer and stretched
across it, in an apparently random vertical position.

Now we draw them aligned on the left side of the viewer, next to the
condition they are warning you about - for example "Engines Critical"
will be drawn in the viewer just to the right of the Engines guage.
In addition, the text is no longer stretched across the display.  They
are definitely visible, but less "In your face".

The pulse-slow animation was changed to prevent the message(s) from
fading entirely as they pulse.