Redo the way UDP connections are negotiated and maintained
In the old Conquest, UDP support was added in a "hacky" way, with the
primary goal of not breaking the protocol.  You had a 5 second window
at the beginning of the connection to the server, always, in which to
send any packet.  Once a packet was received (regardless of content)
the UDP link was established.

If you were not using UDP (passed -u option to Conquest), then you would
have to wait 5 seconds before entering the menu, just for kicks, while
the server's UDP attempt timed out.

And if it broke later on once established, oh well too bad.

This commit removes all of that hackery and replaces it with a new
protocol request and an actual negotiation between the client and the
server in which a UDP connection can be established (or shut down, if
desired).  In addition, once shut down, for whatever reason, it is now
possible to re-nogiate a UDP link afterward.

In addition, during the negotiation, the client must return the proper
number (the server's PID) back to the server to confirm it's
authenticity in requesting the connection.

By default now, an attempt to make a UDP link will be done when
entering the main menu.

This is all a lot more robust and reliable than the hack used in the
past.

Also in this commit - fix up nMenu so it can react properly to an
idle timeout in the menu.  You are only allowed to sit in the menu
without doing anything for 5 minutes.  After that the server drops the
connection.  Now when this happens, you will at least be told. :)