Change the Menu and Help colors to be more readable
Trying to read dark blue text on a black background (especially with
blending) can be harder than it should be :)  So we change the text of
the menu and help screen options to be white, which is easier to see.

In addition, an error beep is generated in the Conquest main menu when
entering a bad command.