Conquest: remove -m (query metaserver) option and make it the default
In the old days, Conquest was frequently played on the same machine as
the server, because Conquest was not a client/server system.  The
previous default for Conquest if run without arguments, was to try to
connect to a server on localhost, mimicking the original behavior.

This default no longer makes any sense.

Now, querying the metaserver is the default behavior, and the -m
option is no longer needed.

Also, fixed up the man page for Conquest somewhat to reflect current
reality.