Reduce usage of utGrand() to only those areas where it's needed
It's really only needed when you need to compare times from different
processes on the machine (client drivers, conqdriv, conquestd, etc),
and they all need to agree on the current "time".

utGetSecs() computes a time based on the previous midnight and
counting seconds after that.  utGrand is supposed to do the same thing
except for counting milliseconds since midnight.  However when this
was implemented in unix, we did not have access to millisecond
granularity, so utGrand() was basically just utGetSecs * 1000.

utSecs/utGrand is expensive, and for many things, we only need to
count a time *difference* between two events, which does not require
a reference to an absolute time.  In addition, these differences do
not need to be shared with anyone (other processes/CB).

This commit converts those uses to use clbGetMillis() which is faster.

There was a "problem" discovered though.

The issue is that in Conquest, there is a minimum wait time for many
things (beaming, bombing, etc).  In most cases the lack of millisecond
granularity didn't affect anything, since the delays were always some
multiple of seconds anyway.

So far, so good.

BOMBARD_GRAND however is different, in that it is the only constant
with sub-second granularity (333 ms), however remember that utGrand()
(on unix anyway) never provided that.

BOMBARD_GRAND is the minimum time to wait before having the
opportunity to kill an army.  With a 1 second granularity (as in old
conquest), this meant you could at most, kill 1 army per second.

With millisecond granularity though, you now have 3 chances per second
to kill an army instead of 1.

To compensate for this problem, which I clearly did not understand at
the time (early 1990's, sigh), changes were made to 2 of the original
conquest constants, DOUBLE_E_KILLS, and BOMBARD_PROB to compensate for
this discrepancy between bombing speed in the original Conquest and
this version.

Now that bombing can happen 3 times faster (due to the millisecond
resolution of the new clock being used), we need to reset those 2
constants back to their original values so that you can't take 50
armies in a single run, as fun as that was.

This will happen in a seperate commit after this one.

In the end though, you will still notice that bombing goes a lot
faster than it used to (about 2x on average - ballpark), but it is
what the original dictated AFAICT.

At any rate, it will be interesting to see how this affects play
compared to the older, much slower, bombing rate you could achieve in
previous versions of Conquest.