Revert some Conquest constants to their original values based on last commit
DOUBLE_E_KILLS is set to 50 (from the changed value of 40)
BOMBARD_PROB is set to .30 (from the changed value of .60)

Read the previous commit message for the backstory.  We reset these
values since the original issue causing them to be changed to try to
compensate for the slow bombing rate was fixed (by now using a
millisecond accurate clock when bombing).

You will notice that bombing is faster than it was, which is not a bad
thing.  Should speed up the games a little :)

I wish I had access to a VAX with original RATFOR Conquest on it to
compare, but the bombing rate should be the same as on this version
now.

The bombing rate was the only operation that required sub-second
precision, which we did not have at the time.  Everything else should
appear to take the same amount of time (beaming, refit, etc).