Delete armies on vacant ships if the ships team was geno'd
When a GENO occurs, we now check all ships that may be carrying the
vanquished team's armies.  If a ship is discovered carrying those
armies, and the ship is vacant, then the armies are removed (beamed
into space, ha ha).

This is meant to prevent a certain cheat whereby someone grabs their
team's armies at the beginning of a game and then hides his ship, and
then goes VACANT, just in case the team is genocided.  You then simply
reincarnate to your ship and beam your armies down somewhere to avoid
having to do a coup.

This is no longer allowed.

Armies on non-VACANT ships are never touched.