Impose a maximum x and y coordinate of +/- 200 million CU's
This limits the size of the universe to 200 million CU's in each
direction.  This is still pretty huge :)

By limiting the size of the universe in this way we can ensure no
overflows occur in the protocol. 200 MCU's should be plenty of space.

When an object exceeds this limit, it will simply wrap around, as
befitting a universe with positive curvature :)