conqoper: Use a more usefel error message when you lack privs to run conqoper
The original message was:

"Poor cretins such as yourself lack the skills necessary to use this
program."

This wasn't particularly helpful, and someone actually opened a bug
ticket on this in 2010.  Ahem. :)

Now it will tell you that you must either run conqoper as root, or you
must be a member of the conquest group.