One more time reworking utGrand and utGetSecs
So, the code was working fine, until a day of runtime had passed.
Then there were continuous player timeouts in the driver when a new
client connected.

So, now we just take the plunge and forget about trying to compute and
maintain offsets from midnight.  We use Epoch time now (seconds since
the Unix Epoch - Jan 1 1970) as the base.

This requires a new common block (for time_t/int64_t storage), and
other adjustments.

utGrand returns milliseconds since the Epoch (in a uint64_t), and
utGetSecs return seconds since the Epoch (in a time_t).  All of these
will be consistent across all processes on the system, regardless of
when each process is started, or how long they have been running.

Hopefully this will fix the driver timeout issues for good.

One oddity - The special stats page in conqoper now reports "Epoch"
time for drivtime and playtime.  This means these will not be offsets
from midnight as they once were, but instead a count of days, hours,
minutes, and seconds since the Epoch (without caring about the current
timezone).  The timezone is irrelevant for our purposes here.

This is fine, as long as you know what you are looking at.  For
timeout purposes, the only thing that matters is that the two times
are within PLAYER_TIMEOUT and DRIVER_TIMEOUT seconds (300).  Aside
from bugs, this should always work now.  Ha ha, that's what I said a
dozen commits ago, but this time it's for real :)