condef.h: get rid of some bad math overrides and redo the one that remains
Get rid of my_rint().

Convert users of round() to iround() (a round that returns an
integer).  This is what conquest was primarily using round() for, but
round() already exists in the math library (returning a double), so
convert the users of round() to iround().

Get rid of round() define - use the real math round() where needed.

Convert users of around() to round(), and get rid of around().

Also, fix up issues reported with cprintf() based on new printf
attribute check implemented in previous commit.