CQI: add ability for a texture to reference another texture
This commit adds the ability for a texture to reference another
texture in the 'filename' element, and use it's attributes (prescale,
color, flags, etc) as a starting point for further definition
(override).

In most cases, certain attributes of a texture like it's prescale
value, need to be propagated to other textures using it (the file).

In the old scheme of things, everyone wanting to use the star texture
for example, would have to set a prescale of 1.7 so it would look
proper in the texture definitions for their stars.

Why not have the star texture set this prescale value, and everyone
else using the texture (to draw various suns) use the same prescale
"automatically" unless specifically overridden?  Huh?  Why?

Well, now you can.  Specify the reference texture in the filename
element preceeded by an '@' sign, and that reference texture (which
must have already been previously defined) is used as the starting
point for further definition of the texture in question.

With that implemented, went through the shipped texturesrc and
00_plantex.trc files and "made it so".

In all cases in 00_plantex.trc, only references to other textures
(with some changes to color for stars) are needed.  This simplified
the file enormously since we don't need to replicate the original
texture's attributes every time we want to use it.  Also switched
Seritil and Dyneb to use the pluto texture.

Renamed the classm.png texture to earth.png, and the classd.png
texture to mars.png.

For texturesrc, created 'base' textures with proper color and prescale
values for the star and all the planet textures we ship (earth, mars,
jupiter, etc).

Redefined class M and class D textures to just reference the new
@earth and @mars, respectively, all texture attributes intact.

For other planet classes: class A uses @neptune, class O uses @jupiter
and class Z uses @pluto.

Updated conqinit.txt to describe '@' syntax for texture filename
elements.

So much nicer.