Rework teams - use flags, other related fixes
This commit reworks teams somewhat, both in the CB and in the
protocol.

We add a flags member with COUPINFO as the currently supported flag.
coupinfo (an integer bool) is gone and represented by a flag now.

We removed the SPTEAM_FLAGS* flags in the protocol spec, and now we
pass the team flags directly via the protocol.

Replaced a bunch of cases where the first character of the team name
was being looked up, rather than the teamchar member already present
in cbTeams.

Moved relevant TEAM_* defines from conqdef.h into team.h.

Replaced a couple of hardcoded local team character lookup arrays with
a single one defined in team.h (TEAM_PLAYERCHARS[]).