packet handling: clean up and improve error detection
Both the client and server should react much faster when a network
error occurrs.  There was some confusion on what constituted an error
(-1 vs. 0 return val) from low level packet reading to the higher
level packet handling.  Higher level, a 0 implies that there is no
data to read (timeout), whereas in low level reading, a 0 return is an
error.  Make sure the distinction is clearer :)

Hopefully this will be somewhat more robust if the peer disconnects
for some unknown reason.